+++ NWN.Anvil/src/main/API/Ruleset/NwSkill.cs
@@ -1,16 +1,14 @@
+using System.Linq;
 using Anvil.Services;
 using NWN.Native.API;
 
 namespace Anvil.API
 {
   /// <summary>
-  /// A creature/character skill.
+  /// A creature/character skill definition.
   /// </summary>
   public sealed class NwSkill
   {
-    [Inject]
-    private static RulesetService RulesetService { get; set; }
-
     [Inject]
     private static TlkTable TlkTable { get; set; }
 
@@ -86,16 +84,29 @@ namespace Anvil.API
       get => TlkTable.GetSimpleString((uint)skillInfo.m_nNameStrref);
     }
 
+    /// <summary>
+    /// Gets the associated <see cref="Skill"/> type for this skill.
+    /// </summary>
     public Skill SkillType { get; }
 
+    /// <summary>
+    /// Resolves a <see cref="NwSkill"/> from a skill id.
+    /// </summary>
+    /// <param name="skillId">The id of the skill to resolve.</param>
+    /// <returns>The associated <see cref="NwSkill"/> instance. Null if the skill id is invalid.</returns>
     public static NwSkill FromSkillId(int skillId)
     {
-      return RulesetService.Skills[skillId];
+      return NwRuleset.Skills.ElementAtOrDefault(skillId);
     }
 
+    /// <summary>
+    /// Resolves a <see cref="NwSkill"/> from a <see cref="Anvil.API.Skill"/>.
+    /// </summary>
+    /// <param name="skillType">The skill type to resolve.</param>
+    /// <returns>The associated <see cref="NwSkill"/> instance. Null if the skill type is invalid.</returns>
     public static NwSkill FromSkillType(Skill skillType)
     {
-      return RulesetService.Skills[(int)skillType];
+      return NwRuleset.Skills.ElementAtOrDefault((int)skillType);
     }
   }
 }